Move toggle button source code into its own folder/add examples
