+++ app/ids-button/example.html
@@ -1,4 +1,3 @@
 {{> button.html }}
 {{> disabled-button.html }}
 {{> icon-button.html }}
-{{> toggle-button.html }}
