--- app/ids-button/toggle-button.html
@@ -1,11 +1,0 @@
-<ids-layout-grid>
-  <ids-label font-size="12" type="h1">Toggle Buttons</ids-label>
-</ids-layout-grid>
-<ids-layout-grid cols="4" gap="md">
-  <ids-layout-grid-cell>
-    <ids-toggle-button id="test-toggle-button" icon-on="star-filled" icon-off="star-outlined" text-off="Toggle Button (Off)" text-on="Toggle Button (On)">
-      <ids-icon slot="icon" icon="settings"></ids-icon>
-      <span slot="text"></span>
-    </ids-toggle-button>
-  </ids-layout-grid-cell>
-</ids-layout-grid>
