+++ app/ids-toggle-button/index.html
@@ -1,0 +1,3 @@
+{{> ../layouts/head.html }}
+{{> example.html }}
+{{> ../layouts/footer.html }}
