+++ app/index.html
@@ -1,5 +1,6 @@
 {{> layouts/head.html title=htmlWebpackPlugin.options.title }}
 {{> ids-button/example.html }}
+{{> ids-toggle-button/example.html }}
 {{> ids-tag/example.html }}
 {{> ids-label/example.html }}
 {{> ids-icon/example.html }}
