+++ webpack.config.js
@@ -20,7 +20,8 @@ module.exports = {
     'ids-popup/ids-popup': ['./app/ids-popup/index.js'],
     'ids-popup/test-sandbox': ['./app/ids-popup/test-sandbox.js'],
     'ids-popup/test-target-in-grid': ['./app/ids-popup/test-target-in-grid.js'],
-    'ids-tag/ids-tag': ['./app/ids-tag/index.js']
+    'ids-tag/ids-tag': ['./app/ids-tag/index.js'],
+    'ids-toggle-button/ids-toggle-button': ['./app/ids-toggle-button/index.js'],
   },
   devtool: 'cheap-source-map', // try source-map for prod
   mode: 'development',
@@ -137,15 +138,15 @@ module.exports = {
     new HTMLWebpackPlugin({
       template: './app/ids-button/disabled-button.html',
       inject: 'body',
-      filename: 'ids-button/toggle-button',
+      filename: 'ids-button/disabled-button',
       chunks: ['ids-button/ids-button', 'ids-button/disabled-button'],
       title: 'IDS Button Component (disabled state)'
     }),
     new HTMLWebpackPlugin({
-      template: './app/ids-button/toggle-button.html',
+      template: './app/ids-button/index.html',
       inject: 'body',
-      filename: 'ids-button/toggle-button',
-      chunks: ['ids-button/ids-button', 'ids-button/toggle-button'],
+      filename: 'ids-toggle-button/index.html',
+      chunks: ['ids-button/ids-button', 'ids-toggle-button/ids-toggle-button'],
       title: 'IDS Toggle Button Component'
     }),
     new HTMLWebpackPlugin({
